+++ packages/renderer/src/ui/whats-new-view.vue
@@ -34,50 +34,18 @@ onMounted(() => {
       <div class="w-[40rem] h-screen px-3 mx-auto">
         <img class="w-20 mx-auto mb-2" src="../assets/icon.png" />
         <p class="text-center text-2xl font-bold mb-8">
-          What's New in Paperlib 1.9.0
+          What's New in Paperlib 1.9.1
         </p>
-        <p class="mt-10"><b>New Features</b></p>
-        <li>RSS subscription. #50</li>
-        <p class="ml-5">支持从 arXiv 等处 RSS 订阅论文。使用方法见下图。</p>
-        <p class="ml-5 text-red-500">
-          ⚠️ If you are using the Atlas database sync, please make sure you have
-          turned on the dev mode of your cloud database to upload the modified
-          table scheme.
-        </p>
-        <p class="ml-5 text-red-500">
-          ⚠️ 如果您使用 Atlas 云同步数据库功能，请确保在使用此功能前打开数据库
-          scheme 的 dev mode
-          保证数据库表结构的修改能够上传到云端。本地数据库无需操作。详情参阅如何使用云同步的教程。
-        </p>
-        <img
-          class="mx-auto mt-5 mb-8 rounded-md shadow-lg"
-          src="../assets/1.jpg"
-        />
-
-        <li>Support markdown notes #94</li>
-        <p class="ml-5">
-          支持 markdown 格式的笔记。只需在编辑的时候输入 &lt;md&gt; 作为开头。
-        </p>
-
-        <img
-          class="mx-auto mt-5 mb-8 rounded-md shadow-lg"
-          src="../assets/2.jpg"
-        />
-
-        <li>A new notification center.</li>
-        <p class="ml-5">
-          添加了一个通知中心。在这里显示 Paperlib 的一些处理信息以及报错。
-        </p>
-
-        <img
-          class="mx-auto mt-5 mb-8 rounded-md shadow-lg w-96"
-          src="../assets/3.jpg"
-        />
 
         <p class="mt-10"><b>Improvements and Fixed Bugs</b></p>
-        <li>More editable information #113</li>
-        <p class="ml-5">更多的可编辑信息。</p>
-        <li>Fix some bugs.</li>
+        <li>Support more RSS subscription formats. #50</li>
+        <p class="ml-5">支持更多的 RSS 返回格式。</p>
+        <li>Fix: Cannot reopen window on macOS #118</li>
+        <p class="ml-5">修复 macOS 快捷键关闭之后无法打开的 Bug。</p>
+        <li>Fix: Text missing under dark mode #117</li>
+        <p class="ml-5">修复 Tooltip 文字 Bug。</p>
+        <li>Fix: Database initialisation bug on Windows.</li>
+        <p class="ml-5">修复 Windows 下数据库初始化 Bug。</p>
 
         <div
           class="mt-10 mx-auto flex w-60 h-10 bg-accentlight dark:bg-accentdark text-neutral-50 rounded-md shadow-md cursor-pointer"